requirements.txt
astroid==1.6.5
cffi==1.11.5
click==6.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mypy==0.620
packaging==17.1
pluggy==0.6.0
py==1.5.4
pycparser==2.18
pylint==1.9.2
pyparsing==2.2.0
pystache==0.5.4
PyYAML==3.13
six==1.11.0
tox==3.1.2
typed-ast==1.1.0
virtualenv==16.0.0
wrapt==1.10.11
